[ThemeBundle] Moved TranslationFilesFinder up to Translation/Finder/ namespace
